minor change for style and mutation privilege

+++ src/server/graphql/schema/trade_log/resolvers.js
@@ -62,19 +62,23 @@ const CLEARENCE_STOCK_LIST = async(root, args, ctx, info) =>{
 
 
 
-//HOLD_SUM: [TradeLog_Type!]!
-//HOLD_STOCK_LIST:[String!]!
-
 const CLEARENCE_TABLE = async (root, args, ctx, info) =>{
   return await ClearenceTableDB.find({});
 };
 
 const REMOVE_HOLD_TABLE_ITEM = async(root, args, ctx, info) =>{
-    //console.log(ctx.state.user._id,args.id);
-    let result = await HoldTableDB.deleteMany({id: args.id});
+    if ( ctx.state.user.privilegeLevel > 20 ) {
+      let result = await HoldTableDB.deleteMany({id: args.id});
+      return { NUMBER : result.n, 
+        OK: result.ok };
 
-    return { NUMBER : result.n, 
-             OK: result.ok };
+    } else {
+      return {
+        NUMBER : 0, 
+        OK: 0 
+      };
+    }
+    
 };
